+++ resources/compiledJS/script/menu.js
@@ -433,18 +433,20 @@ function trimPlayerObjectForSaveFile(playerObject) {
     trimmed.perks.forEach((perk, index) => {
         trimmed.perks[index] = { id: perk.id, tree: perk.tree, commandsExecuted: perk.commandsExecuted };
     });
-    return trimmed;
+    return Object.assign({}, trimmed);
 }
+let saveMenuScroll = 0;
+document.querySelector(".savesMenu .saves").addEventListener("wheel", (wheel) => saveMenuScroll = wheel.path[1].scrollTop);
 async function gotoSaveMenu(inMainMenu = false, animate = true) {
     var _a, _b, _c, _d, _e, _f, _g;
     hideHover();
     saves = JSON.parse(localStorage.getItem(`DOT_game_saves`)) || [];
-    state.savesOpen = true;
-    if (!inMainMenu)
-        closeGameMenu(true);
     const saveBg = document.querySelector(".savesMenu");
     const savesArea = saveBg.querySelector(".saves");
     const saveNameInput = saveBg.querySelector(".saveName");
+    state.savesOpen = true;
+    if (!inMainMenu)
+        closeGameMenu(true);
     if (inMainMenu) {
         saveNameInput.classList.add("unavailable");
         saveBg.querySelector(".saveGame").classList.add("unavailable");
@@ -476,11 +478,10 @@ async function gotoSaveMenu(inMainMenu = false, animate = true) {
     saveNameInput.value = player.name + "_save";
     saves = saves.sort((x1, x2) => x2.time - x1.time);
     resetIds();
-    await sleep(5);
     let renderedSaves = 1;
     for (let save of saves) {
-        if (renderedSaves < 15)
-            await sleep(100);
+        if (renderedSaves < 10 && animate)
+            await sleep(110);
         const saveContainer = document.createElement("div");
         const saveCanvas = document.createElement("canvas");
         const saveName = document.createElement("p");
@@ -561,7 +562,7 @@ async function gotoSaveMenu(inMainMenu = false, animate = true) {
             saves.splice(save.id, 1);
             resetIds();
             localStorage.setItem("DOT_game_saves", JSON.stringify(saves));
-            gotoSaveMenu(false, false);
+            gotoSaveMenu(inMainMenu, false);
         });
         let renderedPlayer = new PlayerCharacter(Object.assign({}, save.save.player));
         renderedPlayer.updatePerks(true, true);
@@ -595,7 +596,6 @@ async function gotoSaveMenu(inMainMenu = false, animate = true) {
         saveName.append(textSyntax(totalText));
         saveName.style.display = "flex";
         renderPlayerOutOfMap(148, saveCanvas, saveCtx, "center", renderedPlayer);
-        await sleep(5);
         buttonsContainer.append(saveOverwrite, loadGame, deleteGame);
         saveContainer.append(saveCanvas, saveName, buttonsContainer);
         savesArea.append(saveContainer);
@@ -603,6 +603,7 @@ async function gotoSaveMenu(inMainMenu = false, animate = true) {
             tooltip(saveName.querySelector(".warningOutOfDate"), lang["out_of_date"]);
         }
     }
+    savesArea.scrollBy(saveMenuScroll, saveMenuScroll);
 }
 async function closeSaveMenu() {
     const saveBg = document.querySelector(".savesMenu");
@@ -653,6 +654,10 @@ function resetIds() {
         saves[i].id = i;
     }
 }
+function updatePlayerToPreventCrash() {
+    player.updatePerks(true);
+    player.updateAbilities();
+}
 function createNewSaveGame() {
     const saveBg = document.querySelector(".savesMenu");
     const savesArea = saveBg.querySelector(".saves");
@@ -672,6 +677,7 @@ function createNewSaveGame() {
     localStorage.setItem("DOT_game_saves", JSON.stringify(saves));
     localStorage.setItem("DOT_game_settings", JSON.stringify(settings));
     localStorage.setItem("DOT_game_language", JSON.stringify(lang.language_id));
+    updatePlayerToPreventCrash();
     gotoSaveMenu(false, false);
 }
 function saveToFile(input) {