+++ script/ability.ts
@@ -72,7 +72,9 @@ const less_is_better = {
   resistance_penetration: false,
   base_heal: false,
   damage_multiplier: false,
-  use_range: false
+  use_range: false,
+  health_cost: true,
+  health_cost_percentage: true
 };
 
 const possible_stat_modifiers = [
@@ -90,6 +92,9 @@ const possible_stat_modifiers = [
   "mpP",
   "resistAllV",
   "resistAllP",
+  "physicalDefP",
+  "magicalDefP",
+  "elementalDefP",
   "hitChanceV",
   "hitChanceP",
   "evasionV",
@@ -99,7 +104,9 @@ const possible_stat_modifiers = [
   "critChanceP",
   "critDamageP",
   "damageV",
-  "damageP"
+  "damageP",
+  "regenHpP",
+  "regenMpP",
 ];
 
 const possible_modifiers = [
@@ -167,6 +174,7 @@ class Ability {
     let statusModifiers: any = {}; 
     baseAbility.statusesUser?.forEach((str: string) => statusModifiers = {...statusModifiers, ...getAbiStatusModifiers(user, base.id, str)});
     baseAbility.statusesEnemy?.forEach((str: string) => statusModifiers = {...statusModifiers, ...getAbiStatusModifiers(user, base.id, str)});
+    if(baseAbility.id == "finishing_blow") console.log(values);
     if(baseAbility.summon_status) statusModifiers = {...statusModifiers, ...getAbiStatusModifiers(user, base.id, baseAbility.summon_status)};
     this.name = baseAbility.name;
     this.mana_cost = Math.floor((baseAbility.mana_cost + values.mana_cost.value) * values.mana_cost.modif) ?? 0;