+++ script/enemy.ts
@@ -113,7 +113,7 @@ class Enemy extends Character {
         });
       }
       Object.entries(this.damages).forEach((dmg: any) => {
-        this.damages[dmg[0]] = Math.floor(this.damages[dmg[0]] * (1 + this.level / 17)) + 1;
+        this.damages[dmg[0]] = Math.floor(this.damages[dmg[0]] * (1 + this.level / 30)) + 1;
       });
       this["stats"]["hp"] = this.getHpMax();
       this["stats"]["mp"] = this.getMpMax();